[LD] Clean up studio.js, move StudioSection to separate functional export
